refactor energyprofilegraph
