+++ app/frontend/types/components/EnergyProfileGraph.ts
@@ -1,0 +1,20 @@
+export interface EnergyProfileGraphProps {
+    enabled?: boolean;
+}
+
+export interface GraphData {
+    [key: string]: number | string;
+}
+
+export interface GraphMetadata {
+    title: string;
+    unit: string;
+    yLabelText: string;
+    xTickLabels?: string[];
+    properties?: { [key: string]: { [key: string]: string } };
+}
+
+export interface GraphResponse {
+    metaData: GraphMetadata;
+    graphData: GraphData[];
+}
